requirements.txt
assertpy
requests
typeguard
pyyaml
pycryptodome